zmiany z soboty

+++ idonc-parent/idonc-model/src/java/org/lopatka/idonc/model/data/IdoncPart.java
@@ -14,7 +14,6 @@ import javax.persistence.JoinColumn;
 import javax.persistence.JoinTable;
 import javax.persistence.ManyToOne;
 import javax.persistence.OneToMany;
-import javax.persistence.OneToOne;
 
 import org.hibernate.annotations.Cache;
 import org.hibernate.annotations.CacheConcurrencyStrategy;
@@ -33,7 +32,7 @@ public class IdoncPart implements Serializable, Comparable<IdoncPart> {
 	@Id
     @Column(name="ID")
     @GeneratedValue(strategy=GenerationType.AUTO)
-    private int id;
+    private Long id;
 
     @Column(name="NUMBER")
 	private Long number;
@@ -51,50 +50,16 @@ public class IdoncPart implements Serializable, Comparable<IdoncPart> {
     @LazyToOne(LazyToOneOption.FALSE)
     private List<IdoncLongData> longDataList;
 
-//	@OneToMany
-//	@Cascade({CascadeType.ALL})
-//    @JoinTable(
-//        name="PART_RESULTS",
-//        joinColumns={@JoinColumn(name="PART_ID")},
-//        inverseJoinColumns=@JoinColumn(name="RESULT_ID")
-//        )
-//    @LazyToOne(LazyToOneOption.FALSE)
-//    private List<IdoncResult> results;
-
+	//FIXME czy tutaj na pewno unique ?
 	@ManyToOne
 	@JoinColumn(name = "RESULT_ID", unique=true)
 	@Cascade(CascadeType.ALL)
 	private IdoncResult result;
 
-//	@OneToMany
-//    @Cascade({CascadeType.SAVE_UPDATE})
-//    @JoinTable(
-//        name="PART_USERS",
-//        joinColumns={@JoinColumn(name="PART_ID")},
-//        inverseJoinColumns=@JoinColumn(name="USER_ID")
-//    )
-//    private List<IdoncUser> usersProcessing;
-
+	//FIXME czy tutaj na pewno unique ?
 	@ManyToOne
 	@JoinColumn(name="USER_PROCESSING_ID", unique=true)
 	private IdoncUser userProcessing;
-
-//    @OneToMany
-//    @Cascade({CascadeType.SAVE_UPDATE})
-//    @JoinTable(
-//        name="LOCKED_USERS",
-//        joinColumns={@JoinColumn(name="PART_ID")},
-//        inverseJoinColumns=@JoinColumn(name="USER_ID")
-//    )
-//    private List<IdoncUser> lockedUsers;
-
-
-//    @ManyToOne
-//    @Cascade(CascadeType.ALL)
-//    @JoinTable(name="PROJECT_PARTS_TO_PROCESS",
-//    		joinColumns = {@JoinColumn(name="PART_ID")},
-//    		inverseJoinColumns = @JoinColumn(name = "PROJECT_ID"))
-//    @LazyToOne(LazyToOneOption.FALSE)
 
     @ManyToOne
     @JoinColumn(name="PROJECT_ID")
@@ -161,43 +126,12 @@ public class IdoncPart implements Serializable, Comparable<IdoncPart> {
 		this.longDataList = longDataList;
 	}
 
-//	public List<IdoncResult> getResults() {
-//		return results;
-//	}
-//
-//	public void setResults(List<IdoncResult> results) {
-//		this.results = results;
-//	}
-//
-//	public void addResult(IdoncResult result) {
-//		this.results.add(result);
-//	}
 
-//	public List<IdoncUser> getUsersProcessing() {
-//		return usersProcessing;
-//	}
-//
-//	public void setUsersProcessing(List<IdoncUser> usersProcessing) {
-//		this.usersProcessing = usersProcessing;
-//	}
-//
-//	public void addUserProcessing(IdoncUser user) {
-//		this.usersProcessing.add(user);
-//	}
-
-//	public List<IdoncUser> getLockedUsers() {
-//		return lockedUsers;
-//	}
-//
-//	public void setLockedUsers(List<IdoncUser> lockedUsers) {
-//		this.lockedUsers = lockedUsers;
-//	}
-
-	public int getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
